% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_block_address_counts_2021.R
\docType{data}
\name{census_block_address_counts_2021}
\alias{census_block_address_counts_2021}
\alias{al_bl_abc_2021_address_block_counts}
\alias{ak_bl_abc_2021_address_block_counts}
\alias{ar_bl_abc_2021_address_block_counts}
\alias{az_bl_abc_2021_address_block_counts}
\alias{ca_bl_abc_2021_address_block_counts}
\alias{co_bl_abc_2021_address_block_counts}
\alias{ct_bl_abc_2021_address_block_counts}
\alias{dc_bl_abc_2021_address_block_counts}
\alias{de_bl_abc_2021_address_block_counts}
\alias{fl_bl_abc_2021_address_block_counts}
\alias{ga_bl_abc_2021_address_block_counts}
\alias{hi_bl_abc_2021_address_block_counts}
\alias{ia_bl_abc_2021_address_block_counts}
\alias{id_bl_abc_2021_address_block_counts}
\alias{il_bl_abc_2021_address_block_counts}
\alias{in_bl_abc_2021_address_block_counts}
\alias{ks_bl_abc_2021_address_block_counts}
\alias{ky_bl_abc_2021_address_block_counts}
\alias{la_bl_abc_2021_address_block_counts}
\alias{ma_bl_abc_2021_address_block_counts}
\alias{md_bl_abc_2021_address_block_counts}
\alias{me_bl_abc_2021_address_block_counts}
\alias{mi_bl_abc_2021_address_block_counts}
\alias{mn_bl_abc_2021_address_block_counts}
\alias{mo_bl_abc_2021_address_block_counts}
\alias{ms_bl_abc_2021_address_block_counts}
\alias{mt_bl_abc_2021_address_block_counts}
\alias{nc_bl_abc_2021_address_block_counts}
\alias{nd_bl_abc_2021_address_block_counts}
\alias{ne_bl_abc_2021_address_block_counts}
\alias{nh_bl_abc_2021_address_block_counts}
\alias{nj_bl_abc_2021_address_block_counts}
\alias{nm_bl_abc_2021_address_block_counts}
\alias{nv_bl_abc_2021_address_block_counts}
\alias{ny_bl_abc_2021_address_block_counts}
\alias{oh_bl_abc_2021_address_block_counts}
\alias{ok_bl_abc_2021_address_block_counts}
\alias{or_bl_abc_2021_address_block_counts}
\alias{pa_bl_abc_2021_address_block_counts}
\alias{pr_bl_abc_2021_address_block_counts}
\alias{ri_bl_abc_2021_address_block_counts}
\alias{sc_bl_abc_2021_address_block_counts}
\alias{sd_bl_abc_2021_address_block_counts}
\alias{tn_bl_abc_2021_address_block_counts}
\alias{tx_bl_abc_2021_address_block_counts}
\alias{ut_bl_abc_2021_address_block_counts}
\alias{va_bl_abc_2021_address_block_counts}
\alias{vt_bl_abc_2021_address_block_counts}
\alias{wa_bl_abc_2021_address_block_counts}
\alias{wi_bl_abc_2021_address_block_counts}
\alias{wv_bl_abc_2021_address_block_counts}
\alias{wy_bl_abc_2021_address_block_counts}
\title{Address Counts of Housing Units and Group Quarters per Census Block for each U.S. State}
\format{
A tibble with variable rows and 4 variables:
\describe{
\item{geoid}{chr Concatenation of U.S. Census STATE, COUNTY, TRACT, and BLOCK ids (15 digits).}
\item{year}{int Year in which data was collected.}
\item{measure}{chr Name of the measurement variable.  total_housing_units - Total number of addresses in the entity count list identified as housing units or transitory units by data in the Master Address File (MAF).  total_group_quarters - Total number of addresses in the entity count list identified as group quarters by data in the MAF.}
\item{value}{dbl Recorded value of the measurement value.}
}

An object of class \code{data.table} (inherits from \code{data.frame}) with 57138 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 371954 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 272846 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 310890 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1039448 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 280692 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 99854 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 12026 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 40398 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 780134 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 465436 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 29466 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 350400 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 163760 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 739958 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 409138 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 345060 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 265326 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 285750 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 214558 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 167656 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 94278 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 509462 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 397412 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 507266 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 224484 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 176836 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 473278 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 169134 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 238208 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 63898 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 275946 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 214432 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 114820 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 577640 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 552858 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 360310 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 261616 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 673972 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 83976 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 51300 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 293690 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 142768 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 359436 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 1337516 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 142416 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 326984 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 49224 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 316188 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 406120 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 145118 rows and 4 columns.

An object of class \code{data.table} (inherits from \code{data.frame}) with 107540 rows and 4 columns.
}
\source{
\url{https://www.census.gov/geographies/reference-files/2020/geo/2020addcountlisting.html}
}
\usage{
al_bl_abc_2021_address_block_counts

ak_bl_abc_2021_address_block_counts

al_bl_abc_2021_address_block_counts

ar_bl_abc_2021_address_block_counts

az_bl_abc_2021_address_block_counts

ca_bl_abc_2021_address_block_counts

co_bl_abc_2021_address_block_counts

ct_bl_abc_2021_address_block_counts

dc_bl_abc_2021_address_block_counts

de_bl_abc_2021_address_block_counts

fl_bl_abc_2021_address_block_counts

ga_bl_abc_2021_address_block_counts

hi_bl_abc_2021_address_block_counts

ia_bl_abc_2021_address_block_counts

id_bl_abc_2021_address_block_counts

il_bl_abc_2021_address_block_counts

in_bl_abc_2021_address_block_counts

ks_bl_abc_2021_address_block_counts

ky_bl_abc_2021_address_block_counts

la_bl_abc_2021_address_block_counts

ma_bl_abc_2021_address_block_counts

md_bl_abc_2021_address_block_counts

me_bl_abc_2021_address_block_counts

mi_bl_abc_2021_address_block_counts

mn_bl_abc_2021_address_block_counts

mo_bl_abc_2021_address_block_counts

ms_bl_abc_2021_address_block_counts

mt_bl_abc_2021_address_block_counts

nc_bl_abc_2021_address_block_counts

nd_bl_abc_2021_address_block_counts

ne_bl_abc_2021_address_block_counts

nh_bl_abc_2021_address_block_counts

nj_bl_abc_2021_address_block_counts

nm_bl_abc_2021_address_block_counts

nv_bl_abc_2021_address_block_counts

ny_bl_abc_2021_address_block_counts

oh_bl_abc_2021_address_block_counts

ok_bl_abc_2021_address_block_counts

or_bl_abc_2021_address_block_counts

pa_bl_abc_2021_address_block_counts

pr_bl_abc_2021_address_block_counts

ri_bl_abc_2021_address_block_counts

sc_bl_abc_2021_address_block_counts

sd_bl_abc_2021_address_block_counts

tn_bl_abc_2021_address_block_counts

tx_bl_abc_2021_address_block_counts

ut_bl_abc_2021_address_block_counts

va_bl_abc_2021_address_block_counts

vt_bl_abc_2021_address_block_counts

wa_bl_abc_2021_address_block_counts

wi_bl_abc_2021_address_block_counts

wv_bl_abc_2021_address_block_counts

wy_bl_abc_2021_address_block_counts
}
\description{
Includes total housing units (including transitory units) and total group quarters counts, by 2020 census tabulation block.
These housing unit and group quarters counts represent final counts for the 2020 Census.
}
\keyword{datasets}
