% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dc.census.address.block.count-package.R
\docType{package}
\name{dc.census.address.block.count-package}
\alias{dc.census.address.block.count}
\alias{dc.census.address.block.count-package}
\title{dc.census.address.block.count: Data Package with Address Counts per Census Block}
\description{
Provides access to data in the address block count lists provided by the U.S. Cesus Bureau.
}
\author{
\strong{Maintainer}: Aaron Schroeder \email{ads7fg@virginia.edu} (\href{https://orcid.org/0000-0003-4372-2241}{ORCID})

}
\keyword{internal}
