% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/va_bl_abc_2021_address_block_counts.R
\docType{data}
\name{va_bl_abc_2021_address_block_counts}
\alias{va_bl_abc_2021_address_block_counts}
\title{Address Counts per Census Block.}
\format{
A tibble with 163,492 rows and 7 variables:
\describe{
\item{STATE}{chr Two-digit state FIPS code}
\item{COUNTY}{chr Three-digit county FIPS code}
\item{TRACT}{chr The census tract number consists of six digits with an implied decimal between the fourth and fifth digits corresponding to the basic census tract number but with leading zeroes and trailing zeroes for census tracts without a suffix.}
\item{BLOCK}{chr Four-digit census block number. Census blocks are numbered uniquely from 0000 to 9999 within census tract, which nest within state and county. The first digit of the census block number identifies the block group.}
\item{BLOCK_GEOID}{chr denoting monthly price for the 20-25/lb. size class}
\item{BLOCK_GEOID}{chr Concatenation of STATE, COUNTY, TRACT, and BLOCK fields, described above, for use in joining the census block-level feedback data to the census blocks in the partnership shapefiles.}
\item{TOTAL HOUSING UNITS}{num Total number of addresses in the entity count list identified as housing units or transitory units by data in the Master Address File (MAF).}
\item{TOTAL GROUP QUARTERS}{num Total number of addresses in the entity count list identified as group quarters by data in the MAF.}
}
}
\source{
\url{https://www2.census.gov/geo/docs/reference/2020addresscountlist/51_Virginia_AddressBlockCountList_092021.txt}
}
\usage{
va_bl_abc_2021_address_block_counts
}
\description{
Includes total housing units (including transitory units) and total group quarters counts, by 2020 census tabulation block.
These housing unit and group quarters counts represent final counts for the 2020 Census.
}
\keyword{datasets}
